Extract query class for Person.
