+++ opengever/contact/models/query.py
@@ -100,3 +100,13 @@ class OgdsUserParticipationQuery(ParticipationQuery):
         return self.filter_by(ogds_userid=ogds_user.id)
 
 OgdsUserParticipation.query_cls = OgdsUserParticipationQuery
+
+
+class PersonQuery(BaseQuery):
+
+    searchable_fields = ['firstname', 'lastname']
+
+    def get_by_former_contact_id(self, former_contact_id):
+        return self.filter_by(former_contact_id=former_contact_id).first()
+
+Person.query_cls = PersonQuery
